loadChats의 정의를 models.js로 이동
