+++ static/src/js/index.js
@@ -5,7 +5,7 @@
 */
 import {ChatModel} from './chat-model.js'
 import {updateSidebar} from './sidebar.js'
-import {currChat, selectChat, addChat, saveChats} from './models.js'
+import {currChat, selectChat, addChat, saveChats, loadChats} from './models.js'
 
 // index.html에 있는 내가 상호작용해야하는 요소를 미리 찾아둡니다.
 // 사용자가 음성 입력을 하려고 할 때 누를는 마이크 버튼
@@ -31,21 +31,6 @@ const elemBtnNewChat = document.querySelector('.btn-new-chat')
 // 채팅 히스토리 목록
 const elemNavList = document.querySelector('.nav__list');
 
-function loadChats()
-{
-    let ids = JSON.parse(localStorage.getItem("ids"));
-    if (ids == null)
-    {
-        ids = []
-    }
-    for(let i = 0; i < ids.length; ++i )
-    {
-        const loadedChat = new ChatModel();
-        loadedChat.loadFromLocalStorage(ids[i])
-        addChat(loadedChat)
-    }
-}
-
 // 현재 보이고 있는 채팅의 메세지를 지우고, 지정된 채팅(chatModel)을 표시
 document.addEventListener("chatsUpdated", event => {
     elemChatMessages.textContent='';
@@ -58,8 +43,7 @@ document.addEventListener("chatsUpdated", event => {
     elemChatMessages.scrollTop = elemChatMessages.scrollHeight;
 });
 
-// 둘 중 하나만
-loadChats(); // 1. 실제 사용시
+loadChats();
 updateSidebar();
 // 일단 처음 시작때는 첫번째 채팅을 보는 상태로 시작
 // selectChat(chats[0])
